Add creation of the BITBUCKET_BACKUP_DB folder. dB backup fails if does not exist

Approved-by: Wolfgang Kritzinger

+++ database-postgresql.sh
@@ -21,7 +21,8 @@ function prepare_backup_db {
 }
 
 function backup_db {
-    rm -r "${BITBUCKET_BACKUP_DB}"
+    [ -d "${BITBUCKET_BACKUP_DB}" ] && rm -r "${BITBUCKET_BACKUP_DB}"
+    mkdir -p "${BITBUCKET_BACKUP_DB}"
     run pg_dump -U "${POSTGRES_USERNAME}" -h "${POSTGRES_HOST}" --port=${POSTGRES_PORT} ${PG_PARALLEL} -Fd \
         -d "${BITBUCKET_DB}" ${PG_SNAPSHOT_OPT} -f "${BITBUCKET_BACKUP_DB}"
 }
